+++ tests/data_proc/test_dk_param_chain.py
@@ -1,10 +1,20 @@
+"""
+Test storing of processing steps with associated
+parameter chains into DataKeeper.
+
+"""
+
 from dataclasses import dataclass
 from copy import deepcopy
 import os
+from pprint import pprint
 import shutil
 
-from proc_params import ProcStepParams, SpikeTrainParams
-from data_keeper import DataKeeper, DataFormat
+from model_tuner.data_proc import (
+    ProcStepParams,
+    NetSpikesParams,
+    DataKeeper
+)
 
 
 @dataclass(frozen=True)
@@ -22,9 +32,9 @@ class NestedParams3(ProcStepParams):
     par3: NestedParams2 = NestedParams2()
 
 
-dirpath_dk = (r'D:\WORK\Salvador\repo\model_tuner\proto\opt_alg_hpc'
-              r'\data\test_dk_param_chain')
-shutil.rmtree(dirpath_dk, ignore_errors=True)
+dirpath_dk = r'D:\WORK\Salvador\repo\model_tuner\test_data\test_dk_param_chain'
+if os.path.exists(dirpath_dk):
+    shutil.rmtree(dirpath_dk, ignore_errors=True)
 os.makedirs(dirpath_dk, exist_ok=True)
 
 dk = DataKeeper(dirpath_dk)
@@ -36,7 +46,7 @@ data_info['sim_result'] = {
     'data': 'SIM_RESULT_DATA'
 }
 data_info['spikes'] = {
-    'step_params': SpikeTrainParams(),
+    'step_params': NetSpikesParams(),
     'data': 'SPIKE_DATA'
 }
 data_info['proc_step'] = {
@@ -52,3 +62,6 @@ param_chain = {}
 for data_name, info in data_info.items():
     param_chain[data_name] = info['step_params']
     dk.store_data(info['data'], data_name, deepcopy(param_chain))
+
+# Print metadata
+pprint(dk.list_data())
